Create README.md
# allinoneredirect
Super bad link that redirect target to some page, but in the mean while, takes gps, photo, ip and autofill data.
Just put the code in your web server and change the url in the file saveauto.php
Just for educational porpose! Do not use it for criminal activities! Don't take any responsability on that code!